+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.4 ] - 2024/10/23
+
+  * Added debug logging instructions to console.log version information.
+
 ###### [ 1.0.3 ] - 2024/10/23
 
   * Changed all references of "spotifyplus-card.js" to "spotifyplus_card.js" to match github repository name.